Surpress warnings in Base#interpolate when too many interpolation arguments are passed
This is a temporary workaround. We should reimplement #interpolate in C to both get rid of the warnings and speed up various things.

We definitely do not want any warnings here because translators need to be in control of picking up or omitting interpolation arguments in their translations. Example: { :zero => 'No messages', :one => 'One message', :many => '%{count} messages' }