Allow mutually recursive mappings
Where translations are drawn from different sources which may be using
different codes for the same language it can be useful to have mappings
where two codes fallback to each other.

This patch allows that by ensuring that the recursion in the computation
of the mapping terminates instead of running out of stack.