Add I18n.locale_available? and enforce available locales
I18n.locale_available? looks for the passed locale in the `I18n.available_locales` list.

---

Enforce available locales

When `I18n.config.enforce_available_locales` is true we'll raise an
I18n::InvalidLocale exception if the passed locale is available.

The default is set to `nil` which will display a deprecation error.

If set to `false` we'll skip enforcing available locales altogether (old behaviour).

This has been implemented in the following methods :

- I18n.config.default_locale=
- I18n.config.locale=
- I18n.translate
- I18n.localize
- I18n.transliterate