Require with full file path to mimic rake so that tests pass on Ruby 1.8.7
Since the require path is different simple_test file will be loaded
twice in Ruby 1.8.7, and the suite then fails with duplicated test names
being defined.

The solution for now is to require the file with full path, which mimics
how rake requires the files and thus only loads it once. This seems to
work with 1.8.7 and 1.9.3, but can be dropped as soon as 1.8.7 support
is removed.