Bugfix chain lookup
The backend chain has a bug that causes it to look up all backends even if it in fact has already found the translation. The reason is a missing explicit return from that point. This can result in extremely slow lookups if a key can be found in a local backend like the yml file, but has to be looked up in the database backend as well.