Introduce middleware that resets current local thread's config after each request
This is a proposal PR.

Rack based apps that work in a multithreaded environment can plugin this
middleware to make sure a context of `locale` is only persisted
in the lifecycle of the request and aren't leaked between requests unintentionally.
Which is common in thread pool like environments.

This is especially helpful if a server/environment isn't cleaning thread local
storage after/before request.