Handle batch key translations in backend(s)
I18n.translate used to accept an array of keys to translate with a single batch request, but it would immediately execute the code
```ruby
if key.is_a?(Array)
  key.map { |k| backend.translate(locale, k, options) }
else
  backend.translate(locale, key, options)
end
```

Unfortunately, this means that even if your backend is somehow more performant on arrays, your backend doesn't ever receive an array of keys.

Now, instead of the I18n translate interface immediately assuming that a backend will perfrom `translate` serially (one key at a time) I18n just passes the whole list to the backend.

This PR also updates all of the backends (except for cascading backend which seemed particularly complicated to get to play nice with this approach) to support lists of keys.

Obviously this is a backwards-incompatible change. I wonder what you think?