Introduce support for deprecating translation keys
Adds an `I18n::deprecate` API that allows users to mark a key as
deprecated and superseded by another.

`I18n::t` will then use the following logic:

- If the key supersedes a deprecated one, and the old one is still
  defined, use the old one instead. If the old one is not defined
  anymore, use the new one.

- If the key is deprecated and the old one is still defined, use
  the old one. If the old one is not defined anymore, use the new
  one.

- For all other keys, regular behavior is maintained.

The reason the old key is always preferred to the new one when
available is that users may have been using the new key for other
purposes, and we don't want to swap it for the old one without
notice. Instead, users have to manually go in and delete the old
key for the new one to start being used.