Add racc to gemspec
According the release note of Ruby 3.3, racc will become the bundled
gem. This patch adds the gem to i18n.gemspec for the next version of
Ruby.

In addition to this change, I made changes in the GitHub workflow to run
tests in Ruby 3.3 because I got an error in Ruby 3.3 when running tests:
`cannot load such file -- racc/parser`. This error is resolved by adding
racc to gemspec.

I want to make sure this Gem works in Ruby 3.3, so the GitHub workflow
is also updated.